# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildPro Construction Business Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/summary:
    get:
      operationId: getJobSummary
      tags: [jobs]
      summary: Get job with related counts/totals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobSummary"

  # ──────────────────────────────────────
  # MATERIAL LISTS (per job)
  # ──────────────────────────────────────
  /jobs/{jobId}/material-list:
    get:
      operationId: getJobMaterialList
      tags: [materials]
      summary: Get (or create) the material list for a job, with items and totals
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialListDetail"
    patch:
      operationId: updateJobMaterialList
      tags: [materials]
      summary: Update material list meta (name/notes)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialListUpdate"
      responses:
        "200":
          description: Updated material list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialListDetail"
  /jobs/{jobId}/material-list/items:
    post:
      operationId: createMaterialItem
      tags: [materials]
      summary: Add an item to the job material list
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialItemInput"
      responses:
        "201":
          description: Created material item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialItem"
  /material-items/{id}:
    patch:
      operationId: updateMaterialItem
      tags: [materials]
      summary: Update a material item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialItemUpdate"
      responses:
        "200":
          description: Updated material item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialItem"
    delete:
      operationId: deleteMaterialItem
      tags: [materials]
      summary: Delete a material item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────
  # ESTIMATES
  # ──────────────────────────────────────
...
  /jobs/{jobId}/photos:
    get:
      operationId: listJobPhotos
      tags: [photos]
      summary: List photos for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobPhoto"
    post:
      operationId: createJobPhoto
      tags: [photos]
      summary: Add a photo to a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPhotoInput"
      responses:
        "201":
          description: Created photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPhoto"
  /job-photos/{id}:
    patch:
      operationId: updateJobPhoto
      tags: [photos]
      summary: Update a job photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPhotoUpdate"
      responses:
        "200":
          description: Updated photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPhoto"
    delete:
      operationId: deleteJobPhoto
      tags: [photos]
      summary: Delete a job photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────
  # RECEIPTS
  # ──────────────────────────────────────
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: [integer, "null"]
        clientName:
          type: [string, "null"]
        status:
          type: string
        jobType:
          type: [string, "null"]
        priority:
          type: string
        description:
          type: [string, "null"]
        address:
          type: [string, "null"]
        city:
          type: [string, "null"]
        state:
          type: [string, "null"]
        zipCode:
          type: [string, "null"]
        startDate:
          type: [string, "null"]
        endDate:
          type: [string, "null"]
        estimatedValue:
          type: [number, "null"]
        actualCost:
          type: [number, "null"]
        notes:
          type: [string, "null"]
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        clientId:
          type: [integer, "null"]
        status:
          type: string
        jobType:
          type: [string, "null"]
        priority:
          type: string
        description:
          type: [string, "null"]
        address:
          type: [string, "null"]
        city:
          type: [string, "null"]
        state:
          type: [string, "null"]
        zipCode:
          type: [string, "null"]
        startDate:
          type: [string, "null"]
        endDate:
          type: [string, "null"]
        estimatedValue:
          type: [number, string, "null"]
        actualCost:
          type: [number, string, "null"]
        notes:
          type: [string, "null"]
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: [integer, "null"]
        status:
          type: string
        jobType:
          type: [string, "null"]
        priority:
          type: string
        description:
          type: [string, "null"]
        address:
          type: [string, "null"]
        city:
          type: [string, "null"]
        state:
          type: [string, "null"]
        zipCode:
          type: [string, "null"]
        startDate:
          type: [string, "null"]
        endDate:
          type: [string, "null"]
        estimatedValue:
          type: [number, string, "null"]
        actualCost:
          type: [number, string, "null"]
        notes:
          type: [string, "null"]
    JobSummary:
      type: object
      required: [job, photoCount, receiptTotal, estimateCount, hasMaterialList]
      properties:
        job:
          $ref: "#/components/schemas/Job"
        photoCount:
          type: integer
        receiptTotal:
          type: number
        estimateCount:
          type: integer
        hasMaterialList:
          type: boolean
    MaterialItem:
      type: object
      required: [id, materialListId, name, quantity, unit, unitPrice, lineTotal, taxable, markup, sortOrder]
      properties:
        id:
          type: integer
        materialListId:
          type: integer
        name:
          type: string
        description:
          type: [string, "null"]
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        lineTotal:
          type: number
        supplier:
          type: [string, "null"]
        sku:
          type: [string, "null"]
        category:
          type: [string, "null"]
        taxable:
          type: boolean
        markup:
          type: boolean
        notes:
          type: [string, "null"]
        sortOrder:
          type: integer
    MaterialItemInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: [string, "null"]
        quantity:
          type: [number, string]
        unit:
          type: string
        unitPrice:
          type: [number, string]
        supplier:
          type: [string, "null"]
        sku:
          type: [string, "null"]
        category:
          type: [string, "null"]
        taxable:
          type: boolean
        markup:
          type: boolean
        notes:
          type: [string, "null"]
        sortOrder:
          type: integer
    MaterialItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: [string, "null"]
        quantity:
          type: [number, string]
        unit:
          type: string
        unitPrice:
          type: [number, string]
        supplier:
          type: [string, "null"]
        sku:
          type: [string, "null"]
        category:
          type: [string, "null"]
        taxable:
          type: boolean
        markup:
          type: boolean
        notes:
          type: [string, "null"]
        sortOrder:
          type: integer
    MaterialListUpdate:
      type: object
      properties:
        name:
          type: string
        notes:
          type: [string, "null"]
    MaterialListDetail:
      type: object
      required: [id, jobId, name, items, subtotal, taxableSubtotal, markupSubtotal]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        name:
          type: string
        notes:
          type: [string, "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/MaterialItem"
        subtotal:
          type: number
        taxableSubtotal:
          type: number
        markupSubtotal:
          type: number
...
    JobPhoto:
      type: object
      required: [id, jobId, type, imageUrl, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        type:
          type: string
        imageUrl:
          type: string
        caption:
          type: [string, "null"]
        takenAt:
          type: [string, "null"]
        createdAt:
          type: string
    JobPhotoInput:
      type: object
      required: [imageUrl]
      properties:
        type:
          type: string
        imageUrl:
          type: string
        caption:
          type: [string, "null"]
        takenAt:
          type: [string, "null"]
    JobPhotoUpdate:
      type: object
      properties:
        type:
          type: string
        caption:
          type: [string, "null"]
        takenAt:
          type: [string, "null"]
...