# lib/api-zod/src/generated/types/materialListUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BuildPro Construction Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialListUpdate {
  name?: string;
  /** @nullable */
  notes?: string | null;
}